Refactored the test for LIC1 and added additional tests for nonvalid input.
refactor/fix: Refactored the test so that test we exepct to return true are in their own function, test that
should return due to incorrect input are in their own function and tests that fail due to invalid input are in their own test function.
Also have to modify the LIC1 function to handle invalid input.