Refactored the test for LIC2 and added an additional test for nonvalid input
refactor/fix: Refactored the test so that test we expect to return true are in one function,
function that we expect to return false due to incorrect input are in their own function and test that
return false due to invalid info are in their own function. Also had make a change to LIC to make sure it returns
false with invalid EPSILON values.